myproject/requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.1.0
Gunicorn==20.1.0
whitenoise==5.3.0
psycopg2==2.9.2
dj-database-url==0.5.0


